Added additional error handling and removed an uncesseary model
