+++ thinkback/models.py
@@ -21,30 +21,3 @@ class Problem:
 		self.name = name
 		self.desc = desc
 		self.function = function
-		
-class ProblemModule:
-	def __init__(self, file):
-		self.file = file
-		self._allowed_extensions = set(['py'])
-
-	def is_empty(self):
-		if self.file.filename == '':
-			return True
-		return False
-
-
-	def is_allowed(self):
-		return '.' in self.file.filename and \
-			self.file.filename.rsplit('.', 1)[1].lower() in self._allowed_extensions
-
-	def create_file_path(self, problem_id):
-		return os.path.join(app.config['UPLOAD_FOLDER'], ''.join(problem_id))
-
-	def save_files_to_path(self, path, secure_filename):
-		self.file.save(os.path.join(path, secure_filename))
-		self.file.save(os.path.join(path, '__init__.py'))
-
-	def get_file_module(self, module_path, filename):
-		filename = filename.split('.')
-		module = importlib.import_module('.{}'.format(filename[0]), package=module_path)
-		return module
